requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
appdirs==1.4.4
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
audioread==3.0.1
av==10.0.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
bchlib==0.14.0
beautifulsoup4==4.12.2
black==24.10.0
blessed==1.20.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
comm==0.1.3
contourpy==1.1.0
croniter==1.3.14
cycler==0.11.0
Cython==0.29.36
datasets==2.14.2
dateutils==0.6.12
debugpy==1.6.7
decorator==5.1.1
deepdiff==6.3.0
dill==0.3.7
et-xmlfile==1.1.0
evaluate==0.4.0
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.88.0
filelock==3.12.0
fonttools==4.42.0
frozenlist==1.3.3
fsspec==2023.4.0
ftfy==6.3.0
fvcore==0.1.5.post20221221
gdown==4.7.1
h11==0.14.0
h5py==3.12.1
hdbscan==0.8.33
huggingface-hub==0.15.1
idna==3.4
imageio==2.31.1
iniconfig==2.0.0
inquirer==3.1.3
iopath==0.1.10
ipykernel==6.23.3
ipython==8.14.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonlines==3.1.0
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.4
language-tool-python==2.7.1
lazy_loader==0.3
librosa==0.9.1
lightning==2.1.2
lightning-cloud==0.5.33
lightning-utilities==0.8.0
lit==16.0.1
llvmlite==0.40.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==0.4.3
nest-asyncio==1.5.6
networkx==3.1
nlpaug==1.1.11
nltk==3.8.1
noise==1.2.2
numba==0.57.1
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
OpenAttack==2.1.1
opencv-python==4.8.1.78
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
ordered-set==4.1.0
packaging==23.1
pandas==2.0.3
parso==0.8.3
pathspec==0.10.3
pedalboard==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==2.5.2
pluggy==1.0.0
pooch==1.6.0
portalocker==2.7.0
prompt-toolkit==3.0.38
protobuf==4.25.1
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pycparser==2.21
pydantic==1.10.7
pydub==0.25.1
Pygments==2.15.1
PyJWT==2.6.0
pynndescent==0.5.10
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.3.1
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.0.1.post0
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
readchar==4.0.5
regex==2023.6.3
requests==2.28.2
resampy==0.4.2
responses==0.18.0
rich==13.3.4
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.13.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.4.1
soxr==0.3.5
stack-data==0.6.2
starlette==0.22.0
starsessions==1.3.0
strsimpy==0.2.1
sympy==1.11.1
tabulate==0.9.0
tensorboardX==2.6.2.2
termcolor==2.5.0
threadpoolctl==3.1.0
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.0
torchaudio==2.0.1
torchdata==0.6.0
torchmetrics==0.11.4
torchtext==0.15.1
torchvision==0.15.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.30.2
triton==2.0.0
typing_extensions==4.5.0
tzdata==2023.3
umap-learn==0.5.3
urllib3==1.26.15
uvicorn==0.21.1
wcwidth==0.2.6
websocket-client==1.5.1
websockets==11.0.2
wget==3.2
xxhash==3.3.0
yacs==0.1.8
yarl==1.9.1
